+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+Workarea Save For Later 1.1.0 (2019-03-13)
+--------------------------------------------------------------------------------
+
+*   Clean Saved Lists After Predetermined Time Range
+
+    Ensure that `Workarea::SavedList` records are removed after a
+    predetermined length of time (default: 3 months) by way of the [TTL
+    Indexes](https://docs.mongodb.com/manual/core/index-ttl/) feature of
+    MongoDB. Only Saved Lists that do not belong to a User will be cleaned
+    in this manner.
+
+    SAVE4LATER-4
+    Tom Scott
+
+*   Update for workarea v3.4 compatibility
+
+    SAVE4LATER-3
+    Matt Duffy
+
+*   Add details to README
+
+    SAVE4LATER-2
+    Matt Duffy
+
+
+
 Workarea Save For Later 1.0.0 (2018-08-08)
 --------------------------------------------------------------------------------